Part 2 of 3: DetailsList css-in-js conversion - DetailsHeader edition (#5490)
* Revert "Revert "WIP: converted detailsheader""

This reverts commit 783bf8f348ff9dc76b71ace04cf4b52b4ccca870.

* Fixed focus check box on header

* get rid of unused var

* update snapshot and fixing up tests

* change file

* fixing up focus keyboard styling for the borders

* update snapshots

* get rid of spread syntax

* get rid of unused vars
