--- packages/office-ui-fabric-react/src/components/DetailsList/DetailsHeader.scss
@@ -1,304 +1,0 @@
-@import '../../common/common';
-
-$iconForegroundColor: $bodySubtextColor;
-$headerForegroundColor: $bodyTextColor;
-$headerBackgroundColor: $bodyBackgroundColor;
-$dropdownChevronForegroundColor: $ms-color-neutralTertiary;
-$rowHeight: 32px;
-$resizerColor: $ms-color-neutralTertiaryAlt;
-$cellPadding: 8px;
-$isPaddedMargin: 24px;
-
-.root {
-  display: inline-block;
-  background: $headerBackgroundColor;
-  position: relative;
-  min-width: 100%;
-  vertical-align: top;
-  height: $rowHeight;
-  line-height: $rowHeight;
-  white-space: nowrap;
-  box-sizing: content-box;
-  padding-bottom: 1px;
-  border-bottom: 1px solid $bodyDividerColor;
-  cursor: default;
-  user-select: none;
-
-  &.rootIsSelectAllHidden {
-    .cell.cellIsCheck {
-      visibility: hidden;
-    }
-  }
-
-  &:hover .check {
-    opacity: 1;
-  }
-}
-
-.cellWrapperPadded {
-  @include ms-padding-right($isPaddedMargin + $cellPadding);
-}
-
-.cell {
-  @include ms-font-s;
-  color: $headerForegroundColor;
-  @include focus-border();
-  position: relative;
-  display: inline-block;
-  box-sizing: border-box;
-  padding: 0 $cellPadding;
-  border: none;
-  line-height: inherit;
-  margin: 0;
-  height: $rowHeight;
-  vertical-align: top;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  @include ms-text-align(left);
-
-  &.cellIsCheck {
-    position: relative;
-    padding: 0;
-    margin: 0;
-    display: inline-flex;
-    align-items: center;
-
-    &.isSelected {
-      opacity: 1;
-    }
-  }
-
-  &.cellIsActionable {
-    &:hover {
-      color: $bodyTextColor;
-      background: $listHeaderBackgroundHoveredColor;
-    }
-
-    &:active {
-      background: $listHeaderBackgroundPressedColor;
-    }
-  }
-
-  &.cellIsEmpty {
-    text-overflow: clip;
-  }
-
-  &:hover .gripperBarVerticalStyle {
-    display: block;
-  }
-}
-
-.gripperBarVerticalStyle {
-  display: none;
-  position: absolute;
-  @include ms-text-align(left);
-  color: $ms-color-neutralTertiary;
-  @include ms-left(1px);
-}
-
-.cellSizer {
-  @include focus-clear();
-
-  display: inline-block;
-  position: relative;
-  cursor: ew-resize;
-  bottom: 0;
-  top: 0;
-  overflow: hidden;
-  height: inherit;
-  background: transparent;
-  z-index: 1;
-  width: 16px;
-
-  &:after {
-    content: '';
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    width: 1px;
-    background: $resizerColor;
-    opacity: 0;
-    left: 50%;
-  }
-
-  &:focus:after,
-  &:hover:after,
-  &.cellIsResizing:after {
-    opacity: 1;
-    transition: opacity 0.3s linear;
-  }
-
-  &.cellIsResizing:after {
-    @include ms-drop-shadow();
-  }
-}
-
-.cellSizerStart {
-  margin: 0 -8px;
-}
-
-.cellSizerEnd {
-  margin: 0;
-  @include margin-left(-16px);
-}
-
-.collapseButton {
-  text-align: center;
-  transform: rotate(-180deg);
-  transform-origin: 50% 50%;
-  transition: transform 0.1s linear;
-  width: 20px;
-  outline: 0;
-
-  &.collapseButtonIsCollapsed {
-    transform: rotate(0deg);
-  }
-}
-
-.iconOnlyHeader {
-  .nearIcon {
-    @include ms-padding-left(0);
-  }
-}
-
-.nearIcon {
-  color: $iconForegroundColor;
-  opacity: 1;
-  @include ms-padding-left(8px);
-
-  &.sortIcon {
-    @include ms-padding-left(4px);
-    position: relative;
-    top: 1px;
-  }
-}
-
-.cell .filterChevron {
-  color: $dropdownChevronForegroundColor;
-  @include ms-padding-left(4px);
-  vertical-align: middle;
-}
-
-.cellTitle {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: stretch;
-  box-sizing: border-box;
-  overflow: hidden;
-  @include focus-border($position: auto);
-  padding: 0 8px 0 12px;
-}
-
-.cellName {
-  flex: 0 1 auto;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-:global(.ms-TooltipHost).checkTooltip {
-  display: block;
-}
-
-.cellTooltip {
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-}
-
-.sizingOverlay {
-  position: absolute;
-  left: 0;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  cursor: ew-resize;
-  background: rgba(255, 255, 255, 0);
-
-  @include high-contrast {
-    background: transparent;
-    -ms-high-contrast-adjust: none;
-  }
-}
-
-.cell .collapseButton {
-  @include ms-padding-right(0);
-}
-
-.accessibleLabel {
-  @include ms-screenReaderOnly;
-}
-
-.borderWhileDragging {
-  border-style: solid;
-  border-width: 1px;
-  border-color: $ms-color-themePrimary;
-  -webkit-animation: fadeOut 0.2s forwards;
-  animation: fadeOut 0.2s forwards;
-}
-
-.dropHintCircleStyle {
-  display: inline-block;
-  visibility: hidden;
-  position: absolute;
-  bottom: 0;
-  top: -27px;
-  height: 9px;
-  width: 9px;
-  border-radius: 50%;
-  @include ms-margin-left(-5px);
-  top: 34px;
-  overflow: visible;
-  z-index: 10;
-  border: 1px solid $ms-color-themePrimary;
-  background: $ms-color-white;
-}
-
-.dropHintLineStyle {
-  display: inline-block;
-  visibility: hidden;
-  position: absolute;
-  bottom: 0;
-  top: -3px;
-  overflow: hidden;
-  height: 37px;
-  width: 1px;
-  background: $ms-color-themePrimary;
-  z-index: 10;
-}
-
-.dropHintStyle {
-  display: inline-block;
-  position: absolute;
-}
-
-.borderAfterDropping {
-  border-style: solid;
-  border-width: 1px;
-  border-color: $ms-color-themePrimary;
-  -webkit-animation: fadeOut 1.5s forwards;
-  animation: fadeOut 1.5s forwards;
-  @include ms-left(-1px);
-  line-height: 31px;
-}
-
-@-webkit-keyframes fadeOut {
-  from {
-    border-color: $ms-color-themePrimary;
-  }
-  to {
-    border-color: white;
-  }
-}
-
-@keyframes fadeOut {
-  from {
-    border-color: $ms-color-themePrimary;
-  }
-  to {
-    border-color: white;
-  }
-}
