+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsRowCheck.styles.ts
@@ -1,6 +1,5 @@
 import { IDetailsRowCheckStyleProps, IDetailsRowCheckStyles } from './DetailsRowCheck.types';
 import { getGlobalClassNames, getFocusStyle } from '../../Styling';
-import { IsFocusVisibleClassName } from '@uifabric/utilities';
 
 const GlobalClassNames = {
   root: 'ms-DetailsRow-check',
@@ -36,18 +35,12 @@ export const getStyles = (props: IDetailsRowCheckStyleProps): IDetailsRowCheckSt
         backgroundColor: 'transparent',
         border: 'none',
         opacity: 0,
-        height: compact ? 32 : 40,
+        height: compact || isHeader ? 32 : 40,
         width: 40,
         padding: 0,
         margin: 0,
         selectors: {
           '&:hover': {
-            opacity: 1
-          },
-          '&.isHeader': {
-            height: 32
-          },
-          [`.${IsFocusVisibleClassName} &`]: {
             opacity: 1
           }
         }
