+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsRowCheck.tsx
@@ -39,6 +39,7 @@ const DetailsRowCheckBase = (props: IDetailsRowCheckProps) => {
     selected: isPressed,
     anySelected,
     className,
+    isHeader,
     compact
   });
